+++ src/main/kotlin/icu/windea/pls/lang/inspections/script/statement/ScopeCallStatementToExplicitFormInspection.kt
@@ -18,7 +18,7 @@ import icu.windea.pls.script.psi.ParadoxScriptProperty
 import icu.windea.pls.script.psi.ParadoxScriptVisitor
 
 /**
- * 建议将作用域调用转换为显式形式。
+ * 建议将作用域调用转换为普通形式。
  *
  * 检测于文法级别。
  *
@@ -26,13 +26,13 @@ import icu.windea.pls.script.psi.ParadoxScriptVisitor
  *
  * @see ParadoxScopeCallStatementManipulationService
  */
-class ScopeCallStatementToExplicitFormInspection : LocalInspectionTool(), DumbAware {
+class ScopeCallStatementToNormalFormInspection : LocalInspectionTool(), DumbAware {
     override fun buildVisitor(holder: ProblemsHolder, isOnTheFly: Boolean): PsiElementVisitor {
         return object : ParadoxScriptVisitor() {
             override fun visitProperty(element: ParadoxScriptProperty) {
                 ProgressManager.checkCanceled()
-                if (!ParadoxScopeCallStatementManipulationService.canConvertToExplicitForm(element)) return
-                val description = PlsBundle.message("inspection.script.scopeCallStatementToExplicitForm.desc")
+                if (!ParadoxScopeCallStatementManipulationService.canConvertToNormalForm(element)) return
+                val description = PlsBundle.message("inspection.script.scopeCallStatementToNormalForm.desc")
                 val fixes = getFixes(element)
                 holder.registerProblem(element.propertyKey, description, *fixes)
             }
@@ -46,13 +46,13 @@ class ScopeCallStatementToExplicitFormInspection : LocalInspectionTool(), DumbAw
     private class Fix(
         element: PsiElement
     ) : LocalQuickFixAndIntentionActionOnPsiElement(element), IntentionActionWithFixAllOption {
-        override fun getText() = PlsBundle.message("inspection.script.scopeCallStatementToExplicitForm.fix.1.name")
+        override fun getText() = PlsBundle.message("inspection.script.scopeCallStatementToNormalForm.fix.1.name")
 
         override fun getFamilyName() = text
 
         override fun invoke(project: Project, file: PsiFile, editor: Editor?, startElement: PsiElement, endElement: PsiElement) {
             val element = startElement as? ParadoxScriptProperty ?: return
-            ParadoxScopeCallStatementManipulationService.convertToExplicitForm(element, project)
+            ParadoxScopeCallStatementManipulationService.convertToNormalForm(element, project)
         }
     }
 }