+++ src/main/kotlin/icu/windea/pls/lang/intentions/script/ScopeCallStatementToExplicitFormIntention.kt
@@ -11,7 +11,7 @@ import icu.windea.pls.script.psi.ParadoxScriptProperty
 import icu.windea.pls.script.psi.ParadoxScriptPsiService
 
 /**
- * 将作用域调用转换为显式形式。
+ * 将作用域调用转换为普通形式。
  *
  * 检测于文法级别。
  *
@@ -20,18 +20,18 @@ import icu.windea.pls.script.psi.ParadoxScriptPsiService
  * @see ParadoxScopeCallStatementManipulationService
  */
 @Suppress("UnstableApiUsage")
-class ScopeCallStatementToExplicitFormIntention : PsiUpdateModCommandAction<ParadoxScriptProperty>(ParadoxScriptProperty::class.java), DumbAware {
-    override fun getFamilyName() = PlsBundle.message("intention.scopeCallStatementToExplicitForm")
+class ScopeCallStatementToNormalFormIntention : PsiUpdateModCommandAction<ParadoxScriptProperty>(ParadoxScriptProperty::class.java), DumbAware {
+    override fun getFamilyName() = PlsBundle.message("intention.scopeCallStatementToNormalForm")
 
     override fun invoke(context: ActionContext, element: ParadoxScriptProperty, updater: ModPsiUpdater) {
-        return ParadoxScopeCallStatementManipulationService.convertToExplicitForm(element, context.project)
+        return ParadoxScopeCallStatementManipulationService.convertToNormalForm(element, context.project)
     }
 
     override fun isElementApplicable(element: ParadoxScriptProperty, context: ActionContext): Boolean {
         // if property value is a block, caret offset should before or at `{`, so do inline math blocks
         if (!ParadoxScriptPsiService.isBeforeValueLeftBoundEnd(element, context.offset)) return false
 
-        return ParadoxScopeCallStatementManipulationService.canConvertToExplicitForm(element)
+        return ParadoxScopeCallStatementManipulationService.canConvertToNormalForm(element)
     }
 
     override fun stopSearchAt(element: PsiElement, context: ActionContext): Boolean {